Tp v0.2

+++ backend/ms-productos/src/main/java/isi/dan/ms_productos/exception/ProductoNotFoundException.java
@@ -3,7 +3,7 @@ package isi.dan.ms_productos.exception;
 import isi.dan.ms_productos.modelo.Categoria;
 
 public class ProductoNotFoundException extends Exception{
-    public ProductoNotFoundException(Long id){
+    public ProductoNotFoundException(Integer id){
         super("Producto "+id+" no encontrado");
     }
